rename files and fixes lint errors
