+++ pages/imprint.js
@@ -1,4 +1,4 @@
-import FooterComponent from "@/components/Footer/footer";
+import FooterComponent from "@/components/Footer/page-footer";
 import Header from "@/components/Header/page-header";
 import ScrollUp from "@/components/ScrollUpButton/scroll-up-button";
 import { faArrowLeftLong, faEnvelope } from "@fortawesome/free-solid-svg-icons";
