+++ pages/index.js
@@ -1,7 +1,7 @@
 import ArtPiecesPreview from "@/components/ArtPiecesPreview/art-pieces-preview";
 import PressCarousel from "@/components/PressCarousel/press-carousel";
 import CategoryFilter from "@/components/CategoryFilter/category-filter";
-import FooterComponent from "@/components/Footer/footer";
+import FooterComponent from "@/components/Footer/page-footer";
 import GallerySliderPreview from "@/components/GalleryCarouselPreview/gallery-carousel-preview";
 import Header from "@/components/Header/page-header";
 import ScrollUp from "@/components/ScrollUpButton/scroll-up-button";
